feat: 결제 확정 로직 임시 구현하여 첨부

+++ src/main/java/com/bootcamp/paymentdemo/domain/member/dto/SearchMemberResponse.java
@@ -12,7 +12,7 @@ public class SearchMemberResponse {
     private final String email;
     private final String name;
     private final String phone;
-    private final Integer pointBalance;
+    private final Long pointBalance;
 
     public static SearchMemberResponse register(Member member) {
         return new SearchMemberResponse(
