+++ src/main/java/com/bootcamp/paymentdemo/domain/order/entity/ProductOrder.java
@@ -4,7 +4,6 @@ package com.bootcamp.paymentdemo.domain.order.entity;
 import com.bootcamp.paymentdemo.common.entity.Base;
 import com.bootcamp.paymentdemo.domain.product.entity.Product;
 import jakarta.persistence.*;
-import jakarta.persistence.criteria.CriteriaBuilder;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -24,10 +23,10 @@ public class ProductOrder extends Base {
     private String nameSnap;
 
     @Column(nullable = false)
-    private Integer priceSnap;
+    private Long priceSnap;
 
     @Column(nullable = false)
-    private Integer quantity;
+    private Long quantity;
 
     @Column(nullable = false)
     private boolean deleted;
@@ -46,8 +45,8 @@ public class ProductOrder extends Base {
             Product product,
             Order order,
             String nameSnap,
-            Integer priceSnap,
-            Integer quantity
+            Long priceSnap,
+            Long quantity
     ) {
         ProductOrder productOrder = new ProductOrder();
